Add CommandLineQuestionHandler for AskUserQuestionTool
- Single-select: enter a number or custom text
- Multi-select: comma-separated numbers or custom text
- Free-text input as alternative to predefined options
- Update AskUserQuestionTool.md and READMEs
- Refactored example applications to use CommandLineQuestionHandler
